security complete authorization complete

+++ src/main/java/com/mentorme/mentor/service/user/UserService.java
@@ -4,8 +4,6 @@ import java.util.List;
 import com.mentorme.mentor.dto.User.NewUserDto;
 import com.mentorme.mentor.dto.User.UpdateUserDto;
 import com.mentorme.mentor.dto.User.UserDto;
-import com.mentorme.mentor.dto.User.UserSignInDto;
-
 
 public interface UserService {
 
@@ -15,7 +13,7 @@ public interface UserService {
 
     UserDto findById(Long id);
 
-    List<UserDto> findByRoleId(Integer roleId);
+    List<UserDto> findByRoleId(Long roleId);
 
     UserDto update(UpdateUserDto updateUserDto);
